backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.135.1",
    "gitpython>=3.1.46",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
